Clean up #/ imports key validation and rewrite tests
- Restructure validation logic for clarity
- Rewrite fourslash tests using BaselineAutoImports
- Delete investigation notes
