+++ internal/fourslash/tests/autoImportPackageJsonImportsHashSlash_test.go
@@ -1,0 +1,61 @@
+package fourslash_test
+
+import (
+	"testing"
+
+	"github.com/microsoft/typescript-go/internal/fourslash"
+	"github.com/microsoft/typescript-go/internal/testutil"
+)
+
+func TestAutoImportPackageJsonImportsHashSlashNodenext(t *testing.T) {
+	t.Parallel()
+	defer testutil.RecoverAndFail(t, "Panic on fourslash test")
+	const content = `// @Filename: /tsconfig.json
+{
+  "compilerOptions": {
+    "module": "nodenext",
+    "rootDir": "./",
+    "outDir": "build"
+  }
+}
+// @Filename: /package.json
+{
+  "imports": {
+    "#/*": {
+      "types": "./src/*",
+      "default": "./src/*"
+    }
+  }
+}
+// @Filename: /src/domain/entities/entity.ts
+export const entity = 1;
+// @Filename: /src/features/deep/consumer.ts
+entit/**/`
+	f, done := fourslash.NewFourslash(t, nil /*capabilities*/, content)
+	defer done()
+	f.BaselineAutoImportsCompletions(t, []string{""})
+}
+
+func TestAutoImportPackageJsonImportsHashSlashNode16(t *testing.T) {
+	t.Parallel()
+	defer testutil.RecoverAndFail(t, "Panic on fourslash test")
+	const content = `// @Filename: /tsconfig.json
+{
+  "compilerOptions": {
+    "module": "node16"
+  }
+}
+// @Filename: /package.json
+{
+  "imports": {
+    "#/*": "./src/*"
+  }
+}
+// @Filename: /src/domain/entities/entity.ts
+export const entity = 1;
+// @Filename: /src/consumer.ts
+entit/**/`
+	f, done := fourslash.NewFourslash(t, nil /*capabilities*/, content)
+	defer done()
+	f.BaselineAutoImportsCompletions(t, []string{""})
+}
