+++ testdata/baselines/reference/fourslash/autoImports/autoImportPackageJsonImportsHashSlashNode16.baseline.md
@@ -1,0 +1,10 @@
+// === Auto Imports === 
+```ts
+// @FileName: /src/consumer.ts
+entit/**/
+``````ts
+import { entity } from "./domain/entities/entity";
+
+entit
+```
+
