+++ testdata/baselines/reference/fourslash/autoImports/autoImportPackageJsonImportsHashSlashNodenext.baseline.md
@@ -1,0 +1,10 @@
+// === Auto Imports === 
+```ts
+// @FileName: /src/features/deep/consumer.ts
+entit/**/
+``````ts
+import { entity } from "#/domain/entities/entity.js";
+
+entit
+```
+
